@@ -1,4 +1,5 @@
 {
+  "type": "module",
   "name": "webpack-starter",
   "version": "1.0.0",
   "description": "",
@@ -12,7 +13,7 @@
     "dev": "webpack --mode=development",
     "build": "webpack --mode=production",
     "check:prettier": "prettier --check .",
-    "check:eslint": "eslint --ext .ts .",
+    "check:eslint": "eslint .",
     "check:stylelint": "stylelint **/*.{scss,sass}",
     "check:all": "pnpm check:prettier && pnpm check:eslint && pnpm check:stylelint",
     "fix:prettier": "prettier --loglevel warn --write .",
@@ -23,11 +24,10 @@
   "devDependencies": {
     "@types/node": "20.12.7",
     "@types/webpack": "5.28.5",
-    "@typescript-eslint/eslint-plugin": "7.7.1",
-    "@typescript-eslint/parser": "7.7.1",
     "autoprefixer": "10.4.19",
     "css-loader": "7.1.1",
-    "eslint": "8.57.0",
+    "eslint": "9.1.1",
+    "eslint-config-prettier": "9.1.0",
     "mini-css-extract-plugin": "2.9.0",
     "postcss": "8.4.38",
     "postcss-loader": "8.1.1",
@@ -41,6 +41,7 @@
     "ts-loader": "9.5.1",
     "ts-node": "10.9.2",
     "typescript": "5.4.5",
+    "typescript-eslint": "7.1.1",
     "webpack": "5.91.0",
     "webpack-cli": "5.1.4"
   },
